Started transitioning to new HTTP server implementation
Reworking internal RPC

1. Introduced new HTTP server implementation. It is serving XMLRPC
   requests in RPCManager.
2. Moved storage of XMLRPC methods from  XmlRpcServer into XmlRpcMethods structure.
3. Refactored most methods to use XmlRpcMethods structure instead of
   full fledged server.
4. Refactored internal XMLRPC calls on rosmaster. Now they do go
   directly to RPCManager without involvement of any network code.
5. Added debug timings for HttpServerConnection.
6. master now uses a single RPCManager for both serving master requests
   and rosout requests.

Improving timing of HTTP server in both rosmaster and RPCManager

1. Resolved lateny issue with separate sending of HTTP header and body.
   Now both header and body are sent in a single syscall.
2. Eliminated some unnecessary memory allocations in hot loops.
3. Fixed deadlock with recursive mutex in master_link.cpp. It appeaared
   that rosconsole logging is actively requesting cached ROS parameter
   which leads to additional MasterLink request which also has some
   logging methods.

XmlRpcClient now uses new HttpFrame

 - Reworked state machine in XmlRpcManager to use HttpFrame
 - Updated tests for it.

Additionally:
 - Added TCP_NODELAY option to http connection

Moved all http-related files to separate http/ folder

Continued refactoring and improving HTTP server in miniroscore

1. Internals of RPCManager are hidden inside Internal object.
2. Internals of Master are hidden inside Internal object.
3. Reworked handling of HTTP endpoints and filters.
4. Reworked initialization of miniroscore
5. Resolved some of issues in tests for XmlRpcClient.

Implementing front page for rosmaster. It displays information about all
topics and nodes.

Massive work in progress towards multimaster

+++ include/miniros/xmlrpcpp/XmlRpcValue.h
@@ -89,7 +89,7 @@ namespace XmlRpc {
     /*virtual*/ ~XmlRpcValue() { invalidate(); }
 
     /// Creates array of specified size.
-    NODISCARD static XmlRpcValue Array(int size);
+    NODISCARD static XmlRpcValue Array(size_t size);
 
     /// Creates a dictionary object.
     NODISCARD static XmlRpcValue Dict();
